Get collection features

Get collection features given collection_id and optional query parameters (see API specs here).

pgf_collection_features(
  collection_id,
  base_url,
  path,
  filter = NULL,
  limit = 10000,
  offset = 0,
  sortby = NULL,
  bbox = NULL,
  properties = NULL,
  precision = NULL,
  transform = NULL,
  groupby = NULL,
  user = gh_user(),
  verbose = FALSE,
  response = FALSE
)

Arguments

collection_id

A character string of the collection id.

base_url

A character string of the base URL.

path

A character string of the path or endpoint.

filter

A named vector or list of the filter(s) to apply, where the list names correspond to column names and the list values correspond to the desired value, e.g. list(gnis_name = "Sangan River").

limit

A positive whole number that limits the number of features in the response.

offset

A positive whole number that starts the response at an offset.

sortby

A string that sorts the response items by a property. Default is ascending but property name can be prepended with '-' to indicate descending.

bbox

A vector of four numbers indicating bounding box in lon/lat (e.g. c(minLon, minLat, maxLon, maxLat)) that features in response must intersect.

properties

A vector of strings of the column names to include. If NULL (default), all columns are retained.

precision

A positive whole number indicating precision (i.e., number of decimal places) of GeoJSON coordinates.

transform

A character vector with the name of the valid transform function followed by the parameter values (e.g. c("ST_Simplify", 100)).

groupby

A string of the property name (i.e., column name) to group response features by. This would typically be used with an aggregate transform function.

user

A character string of the user. This allows whoever is running the server your are querying to contact you in case of misuse.

verbose

A flag indicating whether to provide verbose GET response.

response

A flag indicating whether to return the raw response object.

Value

A sf object.

See also

Examples

if (FALSE) {
pgf_collection_features("whse_basemapping.fwa_wetlands_poly", limit = 1)
}