Get function result

Call a spatial function and get results.

pgf_function_result(
  function_id,
  base_url,
  path,
  parameters,
  limit = 10000,
  offset = 0,
  sortby = NULL,
  bbox = NULL,
  properties = NULL,
  precision = NULL,
  transform = NULL,
  nocache = NULL,
  user = gh_user(),
  verbose = FALSE,
  response = FALSE
)

Arguments

function_id

A character string of the function id.

base_url

A character string of the base URL.

path

A character string of the path or endpoint.

parameters

A named list of the parameters and values for a specific function.

limit

A positive whole number that limits the number of features in the response.

offset

A positive whole number that starts the response at an offset.

sortby

A string that sorts the response items by a property. Default is ascending but property name can be prepended with '-' to indicate descending.

bbox

A vector of four numbers indicating bounding box in lon/lat (e.g. c(minLon, minLat, maxLon, maxLat)) that features in response must intersect.

properties

A vector of strings of the column names to include. If NULL (default), all columns are retained.

precision

A positive whole number indicating precision (i.e., number of decimal places) of GeoJSON coordinates.

transform

A character vector with the name of the valid transform function followed by the parameter values (e.g. c("ST_Simplify", 100)).

nocache

A character string indicating whether to cache results or not. Possible values include 'true' or 'false' or NULL.

user

A character string of the user. This allows whoever is running the server your are querying to contact you in case of misuse.

verbose

A flag indicating whether to provide verbose GET response.

response

A flag indicating whether to return the raw response object.

Value

A sf object.

See also

Examples

if (FALSE) {
function_id <- "fwa_locatealong"
base_url <- "https://features.hillcrestgeo.ca/"
path <- "fwa"

parameters <- list(
  blue_line_key = 356308001,
  downstream_route_measure = 0
)
pgf_function_result(function_id,
  base_url = base_url,
  path = path,
  parameters = parameters
)
}